requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==0.28.1
PyPDF2==3.0.1
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
